refactor: reorganize hook imports and remove dead code
