--- frontend/src/hooks/index.ts
@@ -1,16 +1,0 @@
-/**
- * Centralized hook exports
- */
-
-export { useConfirmation } from './useConfirmation';
-export { useErrorHandler } from './useErrorHandler';
-export { usePagination } from './usePagination';
-export { useSelection } from './useSelection';
-export { useModal } from './useModal';
-export { useGalleryActions } from './useGalleryActions';
-export { usePublicGallery } from './usePublicGallery';
-export { usePublicSelection } from './usePublicSelection';
-export { useProfileActions } from './useProfileActions';
-export { usePhotoUpload } from './usePhotoUpload';
-export { useErrorDetails } from './useErrorDetails';
-export { useGalleryDragAndDrop } from './useGalleryDragAndDrop';
