+++ frontend/src/hooks/useGalleryActions.ts
@@ -1,13 +1,17 @@
 import { useState, useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { galleryService, type GalleryDetail } from '../services/galleryService';
+import { galleryService } from '../services/galleryService';
 import { photoService } from '../services/photoService';
-import type { PhotoUploadResponse } from '../services/photoService';
-import { shareLinkService, type ShareLink } from '../services/shareLinkService';
-import { useErrorHandler, useConfirmation, useModal } from '../hooks';
+import { shareLinkService } from '../services/shareLinkService';
+import { useConfirmation } from './useConfirmation';
+import { useErrorHandler } from './useErrorHandler';
+import { useModal } from './useModal';
 import { handleApiError } from '../lib/errorHandling';
 import type {
+  GalleryDetail,
   GalleryPhotoSortBy,
+  PhotoUploadResponse,
+  ShareLink,
   ShareLinkCreateRequest,
   ShareLinkUpdateRequest,
   SortOrder,
